move code set related components out of CustomCodes
